week8_assignment/requirements.txt
langchain_core
python-dotenv
streamlit
langchain_community
langserve
bs4
faiss-cpu
beautifulsoup4
wikipedia
langchainhub
sentence_transformers
langchain-huggingface